Support cancellation via Ctrl-C (#145)
The support for signal handling was already hooked up but we didn’t check for cancellation anywhere.

With this change we check for cancellation:

* During JSON Schema loading: for each HTTP request or file open.
* While cleaning the output directory: once per file or directory to be cleaned, also while walking the tree.
* While converting the JSON AST: once per handler.
* While emitting the output files: once per package.

The JSON Schema library doesn’t support cancellation so this is implemented there by supplying a filesystem implementation that supports cancellation; there is no way to supply an implementation for the HTTP requests so instead we add cancellation to the global HTTP transport (in `gen.go`). We should try and push a change upstream to either add cancellation overall or allow overriding the HTTP client/transport.

Fixes #116.