Omit types that aren't referred to from a resource (#170)
* Add StripUnusedDefinitions pass to CodeGenerator.Generate

Update expected output to remove unreferenced types in the expected
output for AllOf_generates_wrapper_type

* Add Referees method to Type and impls

The name is awkward - it might be better to replace References with
this.

* Implement collecting referrers (I think)

* Implement the connectedness check

If a definition isn't connected by a series of references to any root
definition (ie, a resource) then it should be filtered out.

Memoised to save checking a type name twice.

* Correct implementation of TypeName.Referees

Originally implementation was based on a misunderstanding of how the
different astmodel types worked.

* Strip unused definitions in jsonast golden tests

This required moving StripUnusedDefinitions out of pkg/codegen - I put
it into pkg/astmodel because that's where all of the types it
interacts with are. Also enabled passing in a different set of roots,
since the golden tests use the Test type (which isn't a resource) as
the root.

* Update Type.References to return a TypeNameSet, remove Referees

Returning a set makes the old behaviour simple to implement in the few
places that used it.

Adds a SetUnion to combine two TypeNameSets. I thought about making
this a method to add all of the elements of another set to the
receiver - it would be more efficient but I think it's less clean.

* Don't memoise negative checkWithPath results

These can come from a search that is unwinding because it detected a
cycle, rather than one that didn't find a root. In that case we can't
know whether a search from this node with a different path would
succeed.

Added a unit test to explore this problem.

Made the TypeNameSet methods take TypeNames rather than *TypeNames -
this simplified them a bit (at the cost of having to do some extra nil
checks at the edges).

* Corrected grammar on doc comments

* Simplify connectedness-check thanks to a suggestion from Matt

Rather than walking up the references until we find a root, walk down
the references from the roots collecting up all of the types we
reach. This is much simpler and doesn't need any memoisation (since
the accumulated set replaces it).

* Update Function.References to match Type.References

* Remove unnecessary nil checks

These *TypeNames can't be nil - there are a lot of earlier points
in the code that do no-look dereferences.