Add a pipeline stage to check for remaining AnyType instances (#203)
Raises an error with the containing packages (so they can be added to
the filtering) if any are found.

Added exclusions in azure-arm.yaml for all groups/versions that fail
at the moment. This gives us a list of the packages that need to be
worked through.

If you want to see the specific types containing AnyType run with
verbosity level 2 or higher. (Adjusted the type filtering logging
level up to 3 to avoid drowning in skipped types at the same time.)