Filter out AnyType packages in the check stage (#238)
This allows the pipeline stage to check for unnecessarily excluded packages so we know to remove a filter when we fix a case that generates AnyTypes in the output. (For example, a PR that fixed the AnyType in `microsoft.importexport/v20161101` was landed while this change was being made.)

Move the list of packages with AnyTypes out of the filters and into a separate configuration item.