Improve OneOf property naming (#276)
When producing an object for a `OneOf` type, naming information is lost when collapsing types like `AllOf` into objects. To fix this we will produce names before collapsing types in order to preserve the original names.

If there are multiple names produced for an `AllOf` we prefer the one that is derived from a `TypeName` or is at a lower “inheritance” depth.