Deepcopy maps, arrays and optional fields when converting to/from ARM types (#287)
* Foo.name -> Foo.fooName to avoid confusion with FakeResource.name

I was getting very confused by what I thought was FakeResource.name
being optional, when I hadn't clicked that both FakeResource and Foo
have .name and Foo.name is optional.

* Deepcopy maps, arrays and optional types

This required some special handling in the case of enums. These are
handled by adding a flag to the params that suppresses conversion for
TypeNames when the type names are the same, since the only thing that
will be true for is enums - every other type will have an arm variant
generated..

* Removed unnecessary if statement from map copying in ConvertToArm

Since we create the target map before the check and iterating over a
nil map does the right thing, the `if sourceMap != nil` check isn't needed.