Replay tests (#311)
Use `go-vcr` to record interactions with ARM and play them back. This solves a few things:

1. Eventually we expect the test suite to be large and it will take a long time to run the entire thing, making it not feasible to run as part of PR validation.
2. On GitHub, PRs from forks cannot use secrets in any way so external contributors will not be able to have full validation run on their PRs. With record-replay we can validate against canned responses and external contributors can also update the cassette files locally using their own Azure subscription.

`envtest` is used to run the operator in-proc so that we can change the ARM client it is using. A separate `envtest` instance is used for each test in order to isolate them from one-another.

All name generation for resources and deployments has been changed to be deterministic, based upon the test (and resource) names. This is needed to successfully recreate identical request/responses. Subscription IDs are all replaced with the null UUID, so that if an external contributor updates the tests it (a) doesn’t leak their Subscription ID and (b) doesn’t result in a big churn across all cassette files. Authorization headers are always dropped from cassette files.