Add tests for two more Storage resource types, and fixes to support (#320)
* Add create/delete tests for the `StorageAccountsBlobService` and `StorageAccountsBlobServicesContainer` resource types. 
* Attempt to make deployment more robust, if we deploy a second time and the deployment already exists. This could happen due to faults or spurious failures in patching resources.
  * TODO: We need to check the existing deployment against spec and delete/recreate it if it is different?
*  Handle authentication configuration errors as early as possible; the `autorest` library does not check this and will give back an authorizer that does not work.
* Added temporary fix for #319.
* Fix bad log message in one of the matchers.
* Only save recording for a test when it is successful. This makes it easier to develop when trying to fix a test.

## Outstanding items

We need to think about how to handle resources like `StorageAccountsBlobService` better. It _must_ be named `"default"` to work but this is not enforced at the moment.

One idea is that we could merge this into the `StorageAccount` resource type so that it can be configured there, and the controller takes care of creating it separately.

Another issue is that the name is not going to be unique in Kubernetes. For resources with an owner we might need to scope the name under the owner name, but I’m not sure that this is possible; the object is created in Kubernetes before we can touch the name. The user can set AzureName to "default" and Name to something else but this is not obvious.

I opened an issue for this here: https://github.com/Azure/k8s-infra/issues/324