Refer to a specific ASO image tag in the helm chart
Extend the helm-chart-manifests target to get the latest release from
github and include it in the chart values.yaml so that deploying the
chart will get the matching image from the registry.

This will prevent errors caused by the helm chart referring to :latest
but not including CRDs for the types included in the released operator
image (if there is a delay between image release and the chart being
updated).