Continue reconciling Azure SQL managed users for unknown errors (#1336)
These could be setup issues like the AAD admin not being set on the
server - giving up immediately means the only way to get the user
created successfully is to edit the resource to remove the finalizer,
delete it, then re-add. The reconcile loop for regular users doesn't
use this pattern of returning true, nil when it can't reach the server.