Add ErroredType (#368)
`ErroredType` lets us represent errors as associated with a particular type. This will (in future) allow us to attach fatal errors to types, and then _not_ fail the generation if those types are pruned from the output.

As part of this, if an `error` is `panic`’d from the `AsDeclarations` or `AsType` functions it is handled at file generation time; this is now done by `AllOfType` and `OneOfType`.