Move missing status info to a ErroredType warning (#369)
Remove missing status info pipeline stage and put the warning into an `ErroredType`. This means it will be emitted if the type is not pruned. 

Various other fixes to ensure `ErroredType` works properly in the rest of the pipeline (at least for this scenario), e.g. `ErroredType` is allowed to contain a nil `Type`.