Change EnsureUserDatabaseRoles to return just an error
And change the reconciliation code for user and aad user to just treat
that as a provisioning failure, rather than saying that it had
succeeded but there were some errors which is just confusing. We still
try to apply changes to all databases even if there is an error for
one of them.

Also other review changes, thanks @matthchr!