Begin to obey Retry-After (#375)
I feel like this is rather messy. Unsure how to improve it at the moment.

The `Applier` now always returns the value of the `Retry-After` header (or a zero duration). It is up to the caller to interpret that, as `Retry-After` can be used in both success (202 Accepted) and error (non-200) situations. 

Also removed the default requeue delay we were using, now relying on the automatic backoff that controller-runtime gives us.