Add MySQLUser and MySQLAADUser v1alpha2 (#1357)
* Add v1alpha2 MySQLUser

This removes DbName from MySQLUser and adds DatabaseRoles to store
per-database permissions. Roles will now only store server-wide
permissions.

Add conversions between v1alpha1 and v1alpha2 versions.

* Add v1alpha2 MySQLAADUser

This removes DBName from MySQLAADUser and adds DatabaseRoles to store
per-database permissions. Roles will now only store server-wide
permissions.

Add conversions between v1alpha1 and v1alpha2 versions.

* Set up conversion webhooks for MySQLUser and MySQLAADUser

* Review feedback, thanks @matthchr!

* Ensure `preserveUnknownFields: false` is set in all webhook patches

These were set for all types with version conversions but not the
others (which aren't in use since they are still commented out in
kustomization.yaml). Turning them on in the rest to remove one step in
the process of adding conversion webhooks to types in the future.

This setting is required for conversion to work - it seems like the
only reason it's not set in the patches is that they were generated by
kubebuilder before the setting was mandatory.

* Add provisioning state methods to v1alpha2 ASOStatus

* Reimplement MySQLUser and MySQLAADUser reconciliation with v1alpha2

They now check server-level (in USER_PRIVILEGES) and
database-level (SCHEMA_PRIVILEGES) permissions.

* Update controller tests to work with v1alpha2 MySQLUser

* Move system database constant to mysql.SystemDatabase

Also rename the ServerPort and DriverName constants so they don't
repeat the name of the package.

* Change EnsureUserDatabaseRoles to return just an error

And change the reconciliation code for user and aad user to just treat
that as a provisioning failure, rather than saying that it had
succeeded but there were some errors which is just confusing. We still
try to apply changes to all databases even if there is an error for
one of them.

Also other review changes, thanks @matthchr!