RedisCache - ensure FailedProvisioning=false when Provisioned=true (#1395)
* Ensure FailedProvisioning=false when Provisioned=true

A user reported a case where a RedisCache was provisioned and working
but was marked as FailedProvisioning=true. This seems to have happened
because it initially failed (possibly a temporary error that we aren't
checking for), and then on a subsequent periodic reconciliation the
ARM query reported that it was deployed successfully. Use the
SetProvisioned helper method to prevent this state.