mysqlserver: Keep track of the polling URL when creating (#1407)
* Keep track of the polling URL when creating a MySQL server

Previously this was being dropped on the floor unless the create call
returned an AsyncOpIncompleteError. This works fine except in the
case of creating a replica at the same time as the leader server (or
very quickly after). In that case the create server call returns no
error, but querying the poll URL shows that the creation failed
because the SourceServerID didn't match anything. Without capturing
the poll URL on the initial creation the reconciliation gets stuck
because it thinks that the 404 it gets when checking for the
server indicates that it's just not ready yet - provisioning never goes
back to false and it doesn't retry the creation.

* Review tweaks, thanks @matthchr!