Add AZURE_TARGET_NAMESPACES to restrict the namespaces the operator watches (#1559)
* Add a target namespaces config, only watch resources therein

* Initial work on target namespace test

* Get target namespace test working in both cases

* More useful logging when creating test RG fails

* Run the no-target-namespaces test in the CI pipeline

This is handled in the same way as the secret naming version setting,
but the more settings we add (some more are on the way), the more
unwieldy it's going to be. We need to come up with a better way of
making different settings testable.

* Rework install- targets so they don't trample go.mod & .sum

Renamed them to install-tools and install-test-tools, since they're
installing binaries used in the build process rather than code
dependencies.

Run the `go get` commands in a temp directory and dummy module so that
they don't update the ASO go.mod and .sum files with dependencies that
our code doesn't actually depend on.

* Use the unfiltered API reader when looking for AAD identities

When target namespaces are set, there's no guarantee that the
operator's namespace is included. The identity finder always needs to
look in the operator namespace so pass it the API reader which
bypasses the filtered cache.

* Review tweaks, thanks @matthchr!