Don't fail on flattening collisions (#1608)
The new VMSS resources do this (a collision on `Type` property). Technically we can handle this case properly but it will require more work.

We can handle the VMSS case because the user can't normally overwrite the top-level `Type` property, so any `Type` property in the CRD must point to the flattened `Type`.