Annotate a resource with the namespace of the operator handling it (#1594)
* Annotate a resource with the namespace of the operator handling it

This will help users to find the operator configured to reconcile a
resource, and log a warning if it looks like multiple operators are
watching the same namespace.

* Refuse to reconcile a resource owned by another operator

This is safer in the case of operators misconfigured to have
overlapping namespaces. We'll be guaranteed that only one operator
will be trying to manage a resource, and the other operator will log
events complaining that it's also supposed to manage the resource but
isn't.