postgresqluser: Close DBs to prevent leaking connections
This is much simpler than a cache of DBs. I started going down that
path but as always it's removing no-longer needed DBs from the cache
that makes it more complicated. For now this will fix the leak and if
we have a problem with opening and closing connections being too slow
we can fix that then.