Ensure DB connections are closed when we've finished with them (#1619)
* postgresqluser: Close DBs to prevent leaking connections

This is much simpler than a cache of DBs. I started going down that
path but as always it's removing no-longer needed DBs from the cache
that makes it more complicated. For now this will fix the leak and if
we have a problem with opening and closing connections being too slow
we can fix that then.

* mysql[aad]user: Close DB when we're finished with it

This is much simpler than trying to use a DB cache for it.

* azuresql: Close DBs when finished with them

In the azuresqlaction, azuresqlmanageduser and azuresqluser
reconcilers.

* Change region used for MySQL test

The ASO CI subscription can't create servers in eastus2 but westus2 is
allowed at the moment.