Use require rather than assert in Azure SQL
Also in the Ensure* helpers.

The assert library doesn't stop the current test if the assertion
fails, which means that the test run always ends up finishing with a
timeout waiting for something that can't happen because some
prerequisite failed.

In general the require model is better for tests. The downside is
that you might need to run the test multiple times to see all the
problems, but it avoids nonsensical situations where you timeout
deleting a resource that you failed to create.