Remove envtest job timeout for now (#1723)
* Remove envtest job timeout for now

It seems like there's a problem in the job that's causing it to be killed by the timeout, but the way the job is killed prevents us from seeing the output which would let us fix the underlying problem in whichever test is failing.

* Use require rather than assert in Azure SQL

Also in the Ensure* helpers.

The assert library doesn't stop the current test if the assertion fails, which means that the test run always ends up finishing with a timeout waiting for something that can't happen because some prerequisite failed.

In general the require model is better for tests. The downside is that you might need to run the test multiple times to see all the problems, but it avoids nonsensical situations where you timeout deleting a resource that you failed to create.

* Increase test FailoverGracePeriod to 60

Previous runs were failing with an error saying that 60 is the minimum. (Not sure whether this is a new constraint?)

* Rewrite Retry as a non-recursive function

It was producing very annoying stack traces if a test timed out.