Improve recording default timeouts and controller backoff timing (#1731)
* Improve recording test default timeouts

In the abscence of a timeout specified on the cmdline, use recording
mode to determine what the timeout should be. When recording, timeout
is longer (15m), and when replaying the timeout is shorter (2m).

* Improve controller exponential backoff and timing

Reduced the maximum wait time (it was >15m).
Polling speed is now variable based on test mode. In replay mode,
polling is fast. In recording mode, polling is slower.