Update helm chart to 1.8.0 (ASO version 1.0.28631) (#1769)
* Update helm chart to 1.8.0 (ASO version 1.0.28631)

* Always render PURGE_...SECRETS and RECOVER_...SECRETS fields

Tweak the pipeline to convert these bools to strings. Always write
them to be explicit - there was a subtle bug lurking because the
default value of RECOVER_SOFT_DELETED_KEYVAULT_SECRETS is true, but
the chart wouldn't ever write out the field if the value was false, so
chart users wouldn't be able to turn the recovery behaviour off.