Update the JSON schema submodule (#1798)
There was a bug in the JSON schema generator previously where multiple
OpenAPI specifications under the same provider (Microsoft.Authorization
for example) would overwrite one another. The ARM JSON schema generator
team solved this problem by manufacturing multiple documents named based
on the OpenAPI specification file/folder name. We must combine all of
these fabricated documents into a single picture of a single API (as
that's what it's actually attempting to represent).

The code to do this is a bit fragile/hacky, although the code that
existed prior to this was also a bit fragile...