Use a standard pattern for most recording tests
The pattern is <service>_<resource>_test.go for the filename, and
Test_<service>_<resource>_CRUD for the test name. In places where
the "common name" of the service and the actual ARM name are different,
I used the common name as it usually reads a bit clearer (AKS instead of
containerservice, CosmosDB instead of DocumentDB).