operator bundle: Remove the cert volume and mount from the pod (#1793)
* Remove the cert volume and mount from the operator bundle deployment

These consume the certificate generated by cert-manager, but OLM will
generate its own and hook them into the pod. This issue was preventing
the operator from deploying correctly on OCP 4.6 (because the secret
referred to in the mount doesn't exist and k8s 1.19 wouldn't start the pod).

* Remove the conversion section from multiversion CRDs in the bundle

This will be fully generated by OLM from the conversion
webhookdefinitions in the CSV when the operator is installed.

* Update previous bundle version in makefile

1.0.28631 is landed in both the operatorhub and openshift bundle
repos.