Update module paths put in generated types to match new location
Also correct the calculation of config.typesOutputModuleSuffix - since
TypesOutputPath is defined to be relative to the output go.mod
location, joining it to the go module path read from go.mod is the
right way to get the full local package prefix.

The use of `filepath.Abs(config.TypesOutputPath)` in `getModuleSuffix`
meant that it was always generating a path under the working
directory, rather than under the go.mod directory.