Operator-per-namespace multitenancy in ASOv2 (#1837)
* Add target namespaces support to v2 ASO
* Add operator-namespace annotation to v2 ASO
* Add operator mode handling to the v2 operator main
* Add per-operator multitenancy settings to the envtest context
* Update build constraints for new version of golang
* Port target namespace and operator-namespace annotation test
* Port operator namespace conflict test to v2
* Add operator mode tests in v2
* Add ForTestWithConfig

This means we don't need separate targets with different env vars set.

* Ensure a cassette is written even if no ARM requests were made

There are some operator-mode tests that expect to make no ARM
requests, but if there's no cassette file for the test then it can't
be run in replay mode without credentials. So we create an empty
cassette after stopping the recorder in that case.