Swagger nondeterminism fixes (#1841)
Fixes #1838.

We currently have two sources of non-determinism in Swagger imports:

1. If a file (say, `network.json`) refers to an external file (say, `common/types.json`), any types referred to in that file will be pulled into the same output package (same group & version) as the originating file. This is on purpose, so that shared types get copied into each output package (some shared files have no version or group). However, it can cause issues if the types in the external file share names that collide with types in the other package.
    - Fix: add the ability to provide namespace for common external files, which will be respected when referring to types from those files. Also add an error/check that will fail if this does happen so we can try to fix it via config.
      - Currently `Microsoft.Common` is specified for files under the  `common-types/resource-management` directory. Note that these types all disappear during generation due to allof/oneof collapsing, so we don’t actually see them in the output.
2. With Swagger there can be multiple `.json` files under each group/version, and they can share the same type names. When we combine them all into one package then types can get overwritten. We must not do this.
   -   Fix: Change to process each `.json` file individually and then merge the results by package. When typename collisions are found, we check if they are structurally identical. If not, they are resolved by prefixing each of the colliding names with the filename it was sourced from.
       - There is still a potential source of non-determinism here ~~to address before merge~~ which is also addressed:
          1. Type `X` (specified in multiple files) refers to type `Y`. `X` collides but is structurally identical, so it is okay to merge.
          2. However, type `Y` collides with others that are not structurally identical.
          3. Thus, all `Y`s get renamed.
          4. Types `X` still collide and are still structurally identical, but when generating the output we will pick an arbitrary exemplar. This means that the specific typename of `Y` that `X` refers to is random. We address this by sorting the files and always picking the first one.

This is a rather big PR as it reworks several core things. However, the output should not change!

## Minutiae

### Ability to specify namespace for a directory

We previously had [overrides included from upstream in our config](https://github.com/Azure/azure-service-operator/blob/20832716956c8f3e3fa5da11cdda6abd075d815b/hack/generator/azure-arm.yaml#L375-L404), but only applied the `Suffix` override. We now apply the `Namespace` override as well. This is what is used to specify the `Microsoft.Common` namespace/group name.

### Testing structural equality

In order to detect structural equality, this expands the `type.Equals` function to take an additional parameter which can override the equality check for specific types (at the moment only `TypeName`). 

To prevent accidentally-not-passing the `overrides` argument, it is required on `type.Equals`. `TypeEquals` has been modified to take it as an optional varags argument. Callsites which invoked `type.Equals` directly have generally been updated to invoke `TypeEquals` instead (unless they are in a context where they have access to an `overrides` argument).

### `astmodel.DiffTypes`

This is added as I found it very useful when debugging. When `AddAllowDuplicates` fails, it will generate a structural diff between the types.