Update docs to use `kubectl apply --server-side=true`
We can't use client-side apply because the VirtualMachines CRD is too
big to fit in the `last-applied-configuration` annotation kubectl
uses, but using `kubectl create` is a bit unpleasant (since it won't
update existing resources).

Also note to wait for cert-manager to be started before installing
ASO - if we don't do this the ASO webhook-service certificate creation
fails because it can't talk to the cert-manager webhook, and the ASO
pod won't start.