Add a DeepCopy to check that source objects aren't modified
Copy the subject before converting or assigning properties to be sure
that the operation isn't inadvertently modifying the source. We saw
what could be a bug in the property assignment generation code that
might not be revealed by the generated tests if the source is also
changed such that the source and destination are equal.