Add support for Redis cache resources (#2008)
* Turn on generation of Redis cache resources

Includes Redis, RedisFirewallRule, RedisLinkedServer and
RedisPatchSchedule.

Convert RedisConfiguration to be a string map - it's a struct with
some defined fields and additionalProperties set. We don't support
roundtripping that with json.

* Add generated files for the Redis resources

* Test Redis cache resources

Turns off the tparallel lint for the test - we want the firewall rule
test to be a non-parallel subtest, otherwise it clobbers the link
setup.

* Make conflicts that say "try again later" retryable

The Redis RP returns a conflict if you try to update a firewall rule
for a redis instance while also setting up a server link, but the
message says to try again later. Change the classification to
retryable in that case.

* Add test recording

* Add samples for Redis resources