Make kubebuilder:validation:Required work (#2126)
* Make (almost) all properties pointers with json:omitempty

This is required in order to make Go JSON serialization behave the way
that we need it to. All types must be optional (regardless of if the
property is required or not) because of how non-optional types (int, string,
MyType, etc) interact with omitempty.
If a field is json:omitempty but its type is not optional (not a ptr) then
the default value for that type will be omitted from the JSON payload. For
example 0 would be omitted for ints. On the other hand if a field is NOT
json:omitempty then the type is always serialized in the payload
which causes issues for kubebuilder:validation:Required (how can we tell
the user didn't specify that value?)

See https://github.com/Azure/azure-service-operator/issues/1999 for more
details.

Note: This doesn't currently make slices and maps ptrs. We may end up
needing to do that as well. If they aren't ptrs (but they are omitempty)
then there isn't a good way to differentiate between unspecified and
empty, which may be a problem for some RP APIs. Doing this is
significantly more horrible than wrapping other types in ptrs so I
haven't done it for now.

* Add test to ensure that kubebuilder:validation:Required works