Make indexers work for envtests non-cached client (#2161)
Our envtests use a non-cached kubeclient for consistency. One problem
with this is that indexing for CRDs is entirely client-side in the
cached-client layer. Since our envtests don't use that, any queries that
attempt to make use of indicies fail.

This adds a custom kubeclient that simulates indicies for our test
client so that we can at least not *break* if an indexed query is made.
Actual end-to-end tests should probably still be written for cases where
use of an index is required... The focus of our tests is on ASO though,
so as long as the clients index feature works we may even be able to
avoid that.