Generate multitenant deployment yaml files automatically (#2188)
The multitenant deployment files can be generated by running the `controller:make-multitenant-files` task. They'll be generated and attached to new releases.

Also adds an integration test to validate that we can create resources in a multitenant-deployed cluster that is run in the PR validation CI.

* Add controller:kind-create-multitenant-cluster task

This applies the generated multitenant files to a new kind cluster and
creates a secret for the tenant operator, so that it's ready for testing.

* Add test for confirming that multitenant mode works

This is a very simple integration test that creates a resource
group (confirming that defaulting webhooks are working in this
configuration) and a vnet (confirming that conversion webhooks are
working as well).