Change ConvertToARM to return a ptr (#2292)
ConvertToARM returning a struct rather than a ptr to a struct originally
came about because the contract in the ARM conversion code generation
is that handlers should return value types. This contract makes sense for
primitive types. It also makes sense when assigning to slices or maps, as
usually there the type is map[string]Foo rather than map[string]*Foo.

The unfortunate result of this pattern though is that the final type
returned by ConvertToARM is a struct rather than a ptr. Until now that hasn't
been a problem, but for RouteTable we need the ability to modify that struct
by assigning a new Properties field to it, if one doesn't already exist.
That isn't possible if a struct is returned and wrapped in an interface like
we are currently doing.

This change rectifies the issue by having ConvertToARM return a pointer.
This impacts both the top level result but also the "inner" results as we
walk the type graph and convert everything.

This has a few consequences:
 - The ConvertToARM code generation now must dereference the returned
   pointer in order to continue meeting the contract of always returning
   value types. This sometimes results in unnatural code that looks like:
     identity := *identityARM.(*BatchAccountIdentityARM)
     result.Identity = &identity
 - The callers of ConvertToARM can now easily modify the returned types.

In an effort to avoid the unnatural * + & combo above, I put a timeboxed effort
towards doing a better job. It turns out that while the contract of the inner
handler returning a struct is awkward in this case, it's NOT awkward in other
cases such as map and slice handling. There's no trivial way to avoid this as
the conversion generation exists now, and given the complexity of that code I'm
worried about making it more complicated to solve this "problem",
so for now we will make do with it as is.